ref: improve module co-location
