+++ examples/erc721/tests/erc721.rs
@@ -2,11 +2,10 @@
 
 use alloy::{
     primitives::{Address, U256},
-    rpc::types::eth::TransactionReceipt,
     sol,
     sol_types::SolConstructor,
 };
-use e2e::{receipt, send, watch, Emits, ErrorExt, User};
+use e2e::{receipt, send, watch, ErrorExt, EventExt, User};
 
 use crate::abi::Erc721;
 
@@ -68,6 +67,7 @@ async fn errors_when_reusing_token_id(alice: User) -> eyre::Result<()> {
     let alice_addr = alice.address();
     let token_id = random_token_id();
     let _ = watch!(contract.mint(alice_addr, token_id))?;
+
     let err = send!(contract.mint(alice_addr, token_id))
         .expect_err("should not mint a token id twice");
     assert!(err.is(Erc721::ERC721InvalidSender { sender: Address::ZERO }));
@@ -84,7 +84,7 @@ async fn transfers(alice: User, bob: User) -> eyre::Result<()> {
     let token_id = random_token_id();
     let _ = watch!(contract.mint(alice_addr, token_id).from(alice_addr))?;
 
-    let receipt: TransactionReceipt = receipt!(contract
+    let receipt = receipt!(contract
         .transferFrom(alice_addr, bob_addr, token_id)
         .from(alice_addr))?;
 
@@ -113,6 +113,7 @@ async fn errors_when_transfer_nonexistent_token(
     let tx = contract
         .transferFrom(alice_addr, bob.address(), token_id)
         .from(alice_addr);
+
     let err = send!(tx).expect_err("should not transfer a non-existent token");
     assert!(err.is(Erc721::ERC721NonexistentToken { tokenId: token_id }));
     Ok(())
@@ -156,6 +157,7 @@ async fn errors_when_transfer_unapproved_token(
     let contract = Erc721::new(contract_addr, &bob.signer);
     let tx =
         contract.transferFrom(alice_addr, bob_addr, token_id).from(bob_addr);
+
     let err = send!(tx).expect_err("should not transfer unapproved token");
     assert!(err.is(Erc721::ERC721InsufficientApproval {
         operator: bob_addr,
